refactor-#29: adjust customer repository and CQRS
