+++ src/customer/commands/handlers/create-customer-profile.handler.ts
@@ -1,8 +1,11 @@
-import { CreateCustomerProfileCommand } from '../impl';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
-import { CustomerRepository } from 'src/customer/repositories/customer.repository';
 import { RpcException } from '@nestjs/microservices';
+
+import { CreateCustomerProfileCommand } from '../impl';
+import { CustomerRepository } from '../../repositories';
+import { CustomerEntity } from '../../entities';
+import { validateDbError } from '../../../database/helpers';
 
 @CommandHandler(CreateCustomerProfileCommand)
 export class CreateCustomerProfileHandler
@@ -12,18 +15,20 @@ export class CreateCustomerProfileHandler
     private readonly customerRepository: CustomerRepository,
   ) {}
 
-  async execute(command: CreateCustomerProfileCommand) {
-    const { first_name, last_name } = command.createCustomerProfileDto;
-    const customer = await this.customerRepository.create();
-
-    customer.first_name = first_name;
-    customer.last_name = last_name;
+  async execute(command: CreateCustomerProfileCommand): Promise<CustomerEntity> {
+    const customer = await this.customerRepository.create({ ...command.createCustomerProfileDto });
 
     try {
-      await customer.save();
+      await this.customerRepository.save(customer);
+
       return customer;
     } catch (error) {
-      throw new RpcException(error);
+      const { code, message } = validateDbError(error.code);
+
+      throw new RpcException({
+        statusCode: code,
+        errorStatus: message,
+      });
     }
   }
 }
