+++ src/customer/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './customer.repository';
